+++ src/OrdersList/OrdersListSearchConfig.js
@@ -2,6 +2,11 @@ import moment from 'moment';
 
 import { DATE_FORMAT } from '@folio/stripes-acq-components';
 
+import {
+  CUSTOM_FIELD_TYPES,
+  FILTERS,
+} from './constants';
+
 const indexes = [
   'metadata.createdDate',
   'dateOrdered',
@@ -19,31 +24,48 @@ const customSearchMap = {
   'dateOrdered': searchByDate,
 };
 
-function getCqlQuery(query, qindex, dateFormat) {
-  return customSearchMap[qindex]?.(query, dateFormat) || `*${query}*`;
+function getCqlQuery(query, qindex, dateFormat, customField) {
+  return customField?.type === CUSTOM_FIELD_TYPES.DATE_PICKER
+    ? searchByDate(query, dateFormat)
+    : customSearchMap[qindex]?.(query, dateFormat) || `*${query}*`;
 }
 
-const getKeywordQuery = (query, dateFormat) => indexes.reduce(
-  (acc, sIndex) => {
-    const cqlQuery = getCqlQuery(query, sIndex, dateFormat);
+const getKeywordQuery = (query, dateFormat, customFields) => {
+  const customFieldIndexes = (customFields || [])
+    .filter((cf) => [
+      CUSTOM_FIELD_TYPES.DATE_PICKER,
+      CUSTOM_FIELD_TYPES.TEXTBOX_SHORT,
+      CUSTOM_FIELD_TYPES.TEXTBOX_LONG,
+    ].includes(cf.type))
+    .map((cf) => `${FILTERS.CUSTOM_FIELDS}.${cf.refId}`);
 
-    if (acc) {
-      return `${acc} or ${sIndex}=="${cqlQuery}"`;
-    } else {
-      return `${sIndex}=="${cqlQuery}"`;
-    }
-  },
-  '',
-);
+  return [...indexes, ...customFieldIndexes].reduce(
+    (acc, sIndex) => {
+      const customField = customFields?.find((cf) => `${FILTERS.CUSTOM_FIELDS}.${cf.refId}` === sIndex);
+      const cqlQuery = getCqlQuery(query, sIndex, dateFormat, customField);
 
-export function makeSearchQuery(dateFormat) {
+      if (acc) {
+        return `${acc} or ${sIndex}=="${cqlQuery}"`;
+      } else {
+        return `${sIndex}=="${cqlQuery}"`;
+      }
+    },
+    '',
+  );
+};
+
+export function makeSearchQuery(dateFormat, customFields) {
   return (query, qindex) => {
     if (qindex) {
-      const cqlQuery = getCqlQuery(query, qindex, dateFormat);
+      const customField = customFields?.find(
+        (cf) => `${FILTERS.CUSTOM_FIELDS}.${cf.refId}` === qindex,
+      );
+
+      const cqlQuery = getCqlQuery(query, qindex, dateFormat, customField);
 
       return `(${qindex}==${cqlQuery})`;
     }
 
-    return getKeywordQuery(query, dateFormat);
+    return getKeywordQuery(query, dateFormat, customFields);
   };
 }