+++ src/OrdersList/hooks/useBuildQuery/useBuildQuery.js
@@ -1,4 +1,7 @@
-import { useCallback } from 'react';
+import {
+  useCallback,
+  useMemo,
+} from 'react';
 
 import {
   makeQueryBuilder,
@@ -10,14 +13,35 @@ import {
 } from '@folio/stripes-acq-components';
 
 import { makeSearchQuery } from '../../OrdersListSearchConfig';
-import { FILTERS } from '../../constants';
+import {
+  CUSTOM_FIELD_TYPES,
+  FILTERS,
+} from '../../constants';
 
-export function useBuildQuery() {
+export function useBuildQuery(customFields) {
   const localeDateFormat = useLocaleDateFormat();
+
+  const customFieldsFilterMap = useMemo(() => {
+    const result = {};
+
+    if (customFields) {
+      customFields.forEach((cf) => {
+        const fieldName = `${FILTERS.CUSTOM_FIELDS}.${cf.refId}`;
+
+        if (cf.type === CUSTOM_FIELD_TYPES.MULTI_SELECT_DROPDOWN) {
+          result[fieldName] = buildArrayFieldQuery.bind(null, fieldName);
+        } else if (cf.type === CUSTOM_FIELD_TYPES.DATE_PICKER) {
+          result[fieldName] = buildDateTimeRangeQuery.bind(null, fieldName);
+        }
+      });
+    }
+
+    return result;
+  }, [customFields]);
 
   return useCallback(makeQueryBuilder(
     'cql.allRecords=1',
-    makeSearchQuery(localeDateFormat),
+    makeSearchQuery(localeDateFormat, customFields),
     'sortby metadata.updatedDate/sort.descending',
     {
       [FILTERS.DATE_CREATED]: buildDateTimeRangeQuery.bind(null, [FILTERS.DATE_CREATED]),
@@ -28,6 +52,8 @@ export function useBuildQuery() {
       },
       [FILTERS.TAGS]: buildArrayFieldQuery.bind(null, [FILTERS.TAGS]),
       [FILTERS.ACQUISITIONS_UNIT]: buildArrayFieldQuery.bind(null, [FILTERS.ACQUISITIONS_UNIT]),
+      ...customFieldsFilterMap,
     },
-  ), [localeDateFormat]);
+  ),
+  [customFieldsFilterMap, localeDateFormat]);
 }